mmr-testing/MMR-MSRDC-MMR-Shim.html

<!DOCTYPE html>
<html>
<head>
    <title>Test Video Element is shimmed</title>
    <meta name="timeout" content="normal">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="mediasource-util.js"></script>
</head>
<body>
    <div id="log"></div>
    <script>
        media_test(function (test) {

            console.log("create element");

            var mediaElement = document.createElement("video");
            if (!document.body) {
                document.body = document.createElement("body");
            }

            console.log("append element");

            document.body.appendChild(mediaElement);

            console.log("check shim");

            assert_true((mediaElement.MSRDC_MMR_Shim != undefined), "Video Element is shimmed");

            console.log(mediaElement.MSRDC_MMR_Shim != undefined);
            
            test.done();
            test.add_cleanup(function () {
                document.body.removeChild(mediaElement);
            });

        }, "Test Video Element is shimmed");

    </script>

</body>
</html>
